generate_operations_zkapp_command
